Removed use of size_t

+++ arch/wasm32/syscall_arch.h
@@ -183,7 +183,7 @@ static __attribute__((always_inline)) long __syscall6(long n, long a, long b, lo
 
 extern long __faasm_read_state(const char *key, unsigned char *buffer, long bufferLen);
 extern void __faasm_write_state(const char *key, unsigned char *data, long dataLen);
-extern void __faasm_write_state_offset(const char *key, size_t offset, unsigned char *data, long dataLen);
+extern void __faasm_write_state_offset(const char *key, long offset, unsigned char *data, long dataLen);
 extern void __faasm_read_state_offset(const char *key, long offset, unsigned char *buffer, long bufferLen);
 extern long __faasm_read_input(unsigned char *buffer, long bufferLen);
 extern void __faasm_write_output(const unsigned char *output, long outputLen);
